Extract rule test helpers to separate file
